0x07/api-rest-calculator/src/test/java/com/example/calculator/controller/CalculatorControllerTest.java
package com.example.calculator.controller;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.RequestBuilder;

import static org.junit.jupiter.api.Assertions.assertEquals;

@ExtendWith(SpringExtension.class)
@WebMvcTest(CalculatorController.class)
public class CalculatorControllerTest {

    @Autowired
    private MockMvc mvc;

    @Test
    void messageWelcome() throws Exception {
        RequestBuilder request = get("/calculator/welcome");
        MvcResult result = mvc.perform(request).andReturn();
        assertEquals("Bem vindo a API Calculator", result.getResponse().getContentAsString());
    }

    @Test
    void addNumbers() throws Exception {
        RequestBuilder request = get("/calculator/addNumbers?number1=5&number2=4");
        MvcResult result = mvc.perform(request).andReturn();
        assertEquals("9.0", result.getResponse().getContentAsString());
    }

    @Test
    void subNumbers() throws Exception {
        RequestBuilder request = get("/calculator/subNumbers?number1=10&number2=5");
        MvcResult result = mvc.perform(request).andReturn();
        assertEquals("5.0", result.getResponse().getContentAsString());
    }

    @Test
    void divideNumbers() throws Exception {
        RequestBuilder request = get("/calculator/divideNumbers?number1=25&number2=5");
        MvcResult result = mvc.perform(request).andReturn();
        assertEquals("5.0", result.getResponse().getContentAsString());
    }

    @Test
    void factorialNumber() throws Exception {
        RequestBuilder request = get("/calculator/factorial?factorial=3");
        MvcResult result = mvc.perform(request).andReturn();
        assertEquals("6", result.getResponse().getContentAsString());
    }

    @Test
    void integerToBinaryNumber () throws Exception {
        RequestBuilder request = get("/calculator/integerToBinary?number1=5");
        MvcResult result = mvc.perform(request).andReturn();
        assertEquals("101", result.getResponse().getContentAsString());
    }

    @Test
    void integerToHexNumber () throws Exception {
        RequestBuilder request = get("/calculator/integerToHexadecimal?number1=789456");
        MvcResult result = mvc.perform(request).andReturn();
        assertEquals("c0bd0", result.getResponse().getContentAsString());
    }

    @Test
    void calculadoraDatas () throws Exception {
        RequestBuilder request = get("/calculator/calculeDayBetweenDate?localDate1=2022-02-04&localDate2=2022-06-26");
        MvcResult result = mvc.perform(request).andReturn();
        assertEquals("142", result.getResponse().getContentAsString());
    }
}